+++ db/schema.rb
@@ -69,7 +69,7 @@ ActiveRecord::Schema[8.0].define(version: 2025_02_28_184900) do
     t.string "name", null: false
     t.string "slogan"
     t.integer "required_experience", default: 0
-    t.float "entry_fee", default: 0.0
+    t.integer "entry_fee", default: 0
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.bigint "specialization_id"
